requirements.txt
acme==0.31.0
asn1crypto==0.22.0
certbot==0.31.0
certbot-apache==0.31.0
certifi==2017.4.17
chardet==3.0.4
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==1.9
Django==2.2.3
django-appconf==1.0.2
django-autoslug==1.9.4
django-blog-zinnia==0.20
django-contact-form==1.5
django-contrib-comments==1.9.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-js-asset==1.1.0
django-mptt==0.9.0
django-recurrence==1.6.0
django-registration==3.0
django-tagging==0.4.6
django-todo==2.4.6
django-xmlrpc==0.1.8
django.js==0.8.1
funcsigs==0.4
future==0.15.2
idna==2.5
josepy==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.0
mock==1.3.0
mysqlclient==1.3.13
ndg-httpsclient==0.4.2
parsedatetime==2.4
pbr==1.8.0
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyICU==1.9.2
pyOpenSSL==17.3.0
pyRFC3339==1.0
python-apt==1.1.0b1+ubuntu0.16.4.5
python-augeas==0.5.0
python-dateutil==2.7.3
pytz==2014.10
requests==2.18.1
requests-toolbelt==0.8.0
six==1.11.0
sqlparse==0.3.0
unattended-upgrades==0.1
urllib3==1.21.1
Whoosh==2.7.4
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
